+++ bigmler/bigmler.py
@@ -52,6 +52,7 @@ from bigmler.analyze.dispatcher import analyze_dispatcher
 from bigmler.cluster.dispatcher import cluster_dispatcher
 from bigmler.anomaly.dispatcher import anomaly_dispatcher
 from bigmler.delete.dispatcher import delete_dispatcher
+from bigmler.report.dispatcher import report_dispatcher
 from bigmler.parser import SUBCOMMANDS
 from bigmler.utils import SYSTEM_ENCODING
 
@@ -96,6 +97,8 @@ def main(args=sys.argv[1:]):
             anomaly_dispatcher(args=new_args)
         elif new_args[0] == "sample":
             sample_dispatcher(args=new_args)
+        elif new_args[0] == "report":
+            report_dispatcher(args=new_args)
         elif new_args[0] == "delete":
             delete_dispatcher(args=new_args)
     else:
@@ -104,6 +107,7 @@ def main(args=sys.argv[1:]):
                  "\n\nbigmler analyze --help\n\n"
                  "\n\nbigmler cluster --help\n\n"
                  "\n\nbigmler anomaly --help\n\n"
+                 "\n\nbigmler report --help\n\n"
                  "\n\nbigmler delete --help\n\n"
                  " for a list of options")